move files testdata + fix lint errors
